LayoutTests/http/tests/site-isolation/history/about-blank-iframe-back-expected.txt

Verifies that an about:blank iframe is handled correctly during back navigation with site isolation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll('iframe').length is 2
PASS event.data.includes('id=cross') is true
PASS successfullyParsed is true

TEST COMPLETE
